requirements.txt
pyjwt==1.5.0
uuid
aniso8601==7.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-GraphQL==2.0.1
Flask-GraphQL-Auth==1.3.2
Flask-SQLAlchemy==2.4.4
graphene==2.1.8
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
promise==2.3
requests==2.25.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
SQLAlchemy==1.3.22
urllib3==1.26.3
Werkzeug==1.0.1

